requirements.txt
alembic
pytest
pytest-asyncio
httpx
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
pydantic
slowapi
pytz
python-dotenv
pydantic[email]
aiosqlite
